- removed some warning - fixed bug in LUT - added direct algo to estimate similarity/isometry and a fast least square estimator for anisotropic isometry estimation - reverted graph interface changes...

+++ devl/nll/matrix-affine-transformation.h
@@ -184,8 +184,8 @@ namespace core
       vector2f spacing;
       for ( ui32 x = 0; x < 2; ++x )
       {
-         spacing[ x ] = std::sqrt( core::sqr( m( 0, x ) ) +
-                                   core::sqr( m( 1, x ) ) );
+         spacing[ x ] = (f32)std::sqrt( core::sqr( m( 0, x ) ) +
+                                        core::sqr( m( 1, x ) ) );
       }
       return spacing;
    }
